requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.21
djangorestframework==3.9.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pymongo==3.8.0
pytz==2019.1
requests==2.22.0
uritemplate==3.0.0
urllib3==1.25.3
